pyproject.toml
[project]
name = "python-business-science"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.1.0",
    "isort>=7.0.0",
    "jedi>=0.19.2",
    "jinja2>=3.1.6",
    "joblib>=1.5.1",
    "jupytext>=1.17.2",
    "matplotlib>=3.10.3",
    "mizani>=0.13.5",
    "mypy>=1.16.0",
    "numpy==1.26.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pandas-flavor>=0.7.0",
    "pandas-stubs>=2.2.3.250527",
    "papermill>=2.6.0",
    "path>=17.1.0",
    "pip>=25.1.1",
    "plotly>=6.1.2",
    "plotnine>=0.14.5",
    "plydata>=0.4.3",
    "pmdarima>=2.0.4",
    "ppscore>=1.1.0",
    "pydantic>=2.11.7",
    "pyjanitor>=0.31.0",
    "radian>=0.6.15",
    "rich>=14.0.0",
    "ruff>=0.14.6",
    "scikit-learn>=1.7.0",
    "scikit-misc>=0.5.1",
    "seaborn>=0.13.2",
    "sktime>=0.38.3",
    "sqlalchemy>=2.0.41",
    "statsmodels>=0.14.4",
    "tqdm>=4.67.1",
    "xlsxwriter>=3.2.3",
    "zict>=3.0.0",
]

[tool.ruff]
# Set the maximum line length to 79.
line-length = 79

[tool.ruff.lint]
# Add the `line-too-long` rule to the enforced rule set. By default, Ruff omits rules that
# overlap with the use of a formatter, like Black, but we can override this behavior by
# explicitly adding the rule.
ignore = ["E402"]


[tool.ruff.lint.pydocstyle]
convention = "google"